refactored some names to maintain a uniform presentation
