+++ src/main/java/com/revature/autosurvey/submissions/controller/ResponseController.java
@@ -1,4 +1,4 @@
-package com.revature.autosurvey.submissionservice.controller;
+package com.revature.autosurvey.submissions.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.revature.autosurvey.submissionservice.service.ResponseService;
+import com.revature.autosurvey.submissions.service.ResponseService;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
