+++ src/main/java/com/revature/autosurvey/submissions/service/ResponseService.java
@@ -1,0 +1,5 @@
+package com.revature.autosurvey.submissions.service;
+
+public class ResponseService {
+
+}
