--- src/main/java/com/revature/autosurvey/submissionservice/service/ResponseService.java
@@ -1,5 +1,0 @@
-package com.revature.autosurvey.submissionservice.service;
-
-public class ResponseService {
-
-}
